complete journal 11 and refactored code
